% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besselK.1.0.R
\name{besselK.1.0}
\alias{besselK.1.0}
\title{Derivative of bessel K wrt nu}
\usage{
besselK.1.0(x, nu, expon.scaled.int = 1, h = 1e-05)
}
\arguments{
\item{x}{Positive numeric.}

\item{nu}{numeric; The order (maybe fractional and negative) of the corresponding Bessel function.}

\item{expon.scaled.int}{integer of either 1 or 2.
If 1, the outputs corresponds to K(x;nu).
If 2, the outputs corresponds to exp(x) K(x;nu) to avoid underflow.}

\item{h}{positive small numeric.}
}
\description{
Derivative of bessel function of the second kind wrt the order variable.
}
\examples{
z <- 0.2
nu <- 0.9
exp(z) * besselK.1.0(z, nu)
besselK.1.0(z, nu, 2)
}
\keyword{besselK}
