% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXr_GIG_smallchi.R
\name{EXr_GIG_smallchi}
\alias{EXr_GIG_smallchi}
\title{E[1/X] of GIG for small chi}
\usage{
EXr_GIG_smallchi(lam, chi, psi, omega = sqrt(chi * psi), eta = sqrt(chi/psi))
}
\arguments{
\item{chi}{non-negative shape parameter. If chi=0, it corresponds to VG}

\item{psi}{non-negative shape parameter. If psi=0, it corresponds to Student's t}

\item{omega}{alternate parametrisation}

\item{eta}{alternate parametrisation}

\item{lambda}{Real shape parameter}
}
\value{
scalar
}
\description{
Expectation of reciprocal moment of GIG random variable
}
\examples{
lambda=0.1
chi=c(1e-20,1e-50,1e-200)
psi=1
EXr_GIG(lambda, chi, psi)
EXr_GIG_smallchi(lambda, chi, psi)
}
\keyword{expectation}
