% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments_GIG_smallchi.R
\name{moments_GIG_smallchi}
\alias{moments_GIG_smallchi}
\title{m-th moment of GIG for small chi}
\usage{
moments_GIG_smallchi(
  m,
  lam,
  chi,
  psi,
  omega = sqrt(chi * psi),
  eta = sqrt(chi/psi)
)
}
\arguments{
\item{m}{Numeric. The moment's order.}

\item{lam}{Real shape parameter \eqn{\lambda}.}

\item{chi}{Non-negative parameter \eqn{\chi}.
Must be positive if \eqn{\lambda \le 0}.
If \eqn{\chi = 0}, it corresponds to the gamma distribution.
Can input a vector for multiple outputs.}

\item{psi}{Non-negative parameter \eqn{\psi}.
Must be positive if \eqn{\lambda \ge 0}.
If \eqn{\psi = 0}, it corresponds to the inverse-gamma distribution.}

\item{omega}{Non-negative scale parameter \eqn{\omega} defined as \eqn{sqrt(\chi * \psi)}.
Can input a vector for multiple outputs (if \code{chi} is not supplied).}

\item{eta}{Non-negative shape parameter \eqn{\eta} defined as \eqn{sqrt(\chi / \psi)}.
Can input a vector for multiple outputs (if \code{chi} is not supplied) with same length as \code{omega}.}
}
\value{
vector with same length as \code{omega}.
}
\description{
m-th moment of GIG random variable for small chi
}
\examples{
m=-1
lam=0.1
chi=c(1e-20,1e-50,1e-200)
psi=1
momentsGIG(m, lam, chi, psi)
momentsGIG(m, lam, 0, psi)
}
\seealso{
Other expectations: 
\code{\link{EX_GIG_smallchi}()},
\code{\link{EX_GIG}()},
\code{\link{EXr_GIG}()},
\code{\link{log_moments_GIG_smallchi}()},
\code{\link{log_moments_GIG}()},
\code{\link{moments_GIG}()}
}
\concept{expectations}
\keyword{expectation}
